+++ internal/downloader/errors.go
@@ -1,0 +1,5 @@
+package downloader
+
+import "errors"
+
+var ErrTargetExists = errors.New("target file already exists")